refactored optimizelyconfig
