+++ OptimizelySDK.Net40/OptimizelySDK.Net40.csproj
@@ -304,8 +304,8 @@
 	<Compile Include="..\OptimizelySDK\OptlyConfig\OptimizelyVariation.cs">
       <Link>OptlyConfig\OptimizelyVariation.cs</Link>
     </Compile>
-	<Compile Include="..\OptimizelySDK\OptlyConfig\IOptimizelyConfig.cs">
-      <Link>OptlyConfig\IOptimizelyConfig.cs</Link>
+	<Compile Include="..\OptimizelySDK\OptlyConfig\OptimizelyConfigService.cs">
+      <Link>OptlyConfig\OptimizelyConfigService.cs</Link>
     </Compile>
   </ItemGroup>
   <ItemGroup>
