+++ OptimizelySDK.NetStandard16/OptimizelySDK.NetStandard16.csproj
@@ -84,7 +84,7 @@
 		<Compile Include="..\OptimizelySDK\OptlyConfig\OptimizelyFeature.cs" />
 		<Compile Include="..\OptimizelySDK\OptlyConfig\OptimizelyVariable.cs" />
 		<Compile Include="..\OptimizelySDK\OptlyConfig\OptimizelyVariation.cs" />
-		<Compile Include="..\OptimizelySDK\OptlyConfig\IOptimizelyConfig.cs" />
+		<Compile Include="..\OptimizelySDK\OptlyConfig\OptimizelyConfigService.cs" />
 		<Compile Include="..\OptimizelySDK\Config\FallbackProjectConfigManager.cs">
 		  <Link>AtomicProjectConfigManager.cs</Link>
 		</Compile>
