+++ OptimizelySDK/OptlyConfig/OptimizelyExperiment.cs
@@ -18,10 +18,8 @@ using System.Collections.Generic;
 
 namespace OptimizelySDK.OptlyConfig
 {
-    public class OptimizelyExperiment
+    public class OptimizelyExperiment : Entity.IdKeyEntity
     {
-        public string ID { get; set; }
-        public string Key { get; set; }
         public Dictionary<string, OptimizelyVariation> VariationsMap = new Dictionary<string, OptimizelyVariation>();
     }
 }
