+++ OptimizelySDK/OptlyConfig/OptimizelyFeature.cs
@@ -17,10 +17,8 @@ using System.Collections.Generic;
 
 namespace OptimizelySDK.OptlyConfig
 {
-    public class OptimizelyFeature
+    public class OptimizelyFeature : Entity.IdKeyEntity
     {
-        public string ID { get; set; }
-        public string Key { get; set; }
         public Dictionary<string, OptimizelyExperiment> ExperimentsMap = new Dictionary<string, OptimizelyExperiment>();
         public Dictionary<string, OptimizelyVariable> VariablesMap = new Dictionary<string, OptimizelyVariable>();
     }
