+++ OptimizelySDK/OptlyConfig/OptimizelyVariable.cs
@@ -16,10 +16,8 @@
 
 namespace OptimizelySDK.OptlyConfig
 {
-    public class OptimizelyVariable
+    public class OptimizelyVariable : Entity.IdKeyEntity
     {
-        public string ID { get; set; }
-        public string Key { get; set; }
         public string Type { get; set; }
         public string Value { get; set; }
     }
